Moving new version of common functions back into common.
